requirements.txt
# Runtime dependencies
werkzeug==1.0.1
Flask==1.1.2
Flask-API==1.1
psycopg2==2.8.4
Flask-SQLAlchemy>=2.5
Flask-Migrate==2.5.2
python-dotenv==0.10.3
psycopg2-binary==2.8.6
gunicorn==20.0.4
honcho>=1.0.1

# Testing
factory-boy==2.12.0
nose==1.3.7
rednose==1.3.0
pinocchio==0.4.2
coverage==4.5.4
pylint>=2.4.4
flake8==3.7.9